add no connection status layout, retry button to fetch data

+++ app/src/main/java/com/example/covidtracker/core/models/CountryData.kt
@@ -9,17 +9,17 @@ import java.io.Serializable
 data class CountryData(
     val active: Long,
     val cases: Long,
-    val casesPerOneMillion: Long,
+    val casesPerOneMillion: Double,
     val continent: String,
     val country: String,
     @Embedded(prefix = "countryInfo")
     val countryInfo: CountryInfo,
     val critical: Long,
     val deaths: Long,
-    val deathsPerOneMillion: Long,
+    val deathsPerOneMillion: Double,
     val recovered: Long,
     val tests: Long,
-    val testsPerOneMillion: Long,
+    val testsPerOneMillion: Double,
     val todayCases: Long,
     val todayDeaths: Long,
     @PrimaryKey
